move string related function to String.hpp
