+++ Etaler/Backends/OpenCLBackend.cpp
@@ -30,24 +30,6 @@ inline size_t selectWorkSize(size_t max, size_t mul_of, size_t size)
 	return std::min((size_t)max, round(size));
 }
 
-inline std::string hash_string(const std::string& str)
-{
-	auto hash = std::hash<std::string>()(str);
-	std::stringstream ss;
-	ss << std::hex << hash;
-	return ss.str();
-}
-
-void replaceAll(std::string& str, const std::string& from, const std::string& to) {
-	if(from.empty())
-		return;
-	size_t start_pos = 0;
-	while((start_pos = str.find(from, start_pos)) != std::string::npos) {
-		str.replace(start_pos, from.length(), to);
-		start_pos += to.length(); // In case 'to' contains 'from', like replacing 'x' with 'yx'
-	}
-}
-
 
 template <typename T>
 std::string str(T&& s)
